--- Mobile-Member/app/src/androidTest/java/com/example/jayclarens/buttomnavigationviewfragmented/ExampleInstrumentedTest.java
@@ -1,26 +1,0 @@
-package com.example.jayclarens.buttomnavigationviewfragmented;
-
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumentation test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() throws Exception {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("com.example.jayclarens.buttomnavigationviewfragmented", appContext.getPackageName());
-    }
-}
