+++ Mobile-Member/app/src/main/java/com/TA/MVP/appmobilemember/FragmentDashboard.java
@@ -1,0 +1,29 @@
+package com.TA.MVP.appmobilemember;
+
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class FragmentDashboard extends Fragment {
+
+
+    public FragmentDashboard() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_dashboard, container, false);
+    }
+
+}
