+++ Mobile-Member/app/src/main/java/com/TA/MVP/appmobilemember/MainActivity.java
@@ -1,0 +1,95 @@
+package com.TA.MVP.appmobilemember;
+
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.design.widget.BottomNavigationView;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v7.app.AppCompatActivity;
+import android.view.MenuItem;
+
+public class MainActivity extends AppCompatActivity {
+    private static final String TAG = MainActivity.class.getSimpleName();
+    private BottomNavigationView bottomNavigation;
+    private Fragment fragment;
+    private FragmentManager fragmentManager;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        bottomNavigation = (BottomNavigationView) findViewById(R.id.bottom_navigation);
+        bottomNavigation.inflateMenu(R.menu.navigation);
+        fragmentManager = getSupportFragmentManager();
+
+        //Untuk inisialisasi fragment pertama kali
+        fragmentManager.beginTransaction().replace(R.id.main_container, new FragmentHome()).commit();
+
+        //Memberikan listener saat menu item di bottom navigation diklik
+        bottomNavigation.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                int id = item.getItemId();
+                switch (id){
+                    case R.id.menu_home:
+                        fragment = new FragmentHome();
+                        break;
+                    case R.id.menu_buku:
+                        fragment = new FragmentBuku();
+                        break;
+                    case R.id.menu_musik:
+                        fragment = new FragmentMusik();
+                        break;
+                    case R.id.menu_film:
+                        fragment = new FragmentFilm();
+                        break;
+                    case R.id.menu_dashboard:
+                        fragment = new FragmentDashboard();
+                        break;
+
+                }
+                final FragmentTransaction transaction = fragmentManager.beginTransaction();
+                transaction.replace(R.id.main_container, fragment).commit();
+                return true;
+            }
+        });
+    }
+}
+//public class MainActivity extends AppCompatActivity {
+//
+//    private TextView mTextMessage;
+//
+//    private BottomNavigationView.OnNavigationItemSelectedListener mOnNavigationItemSelectedListener
+//            = new BottomNavigationView.OnNavigationItemSelectedListener() {
+//
+//        @Override
+//        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+//            switch (item.getItemId()) {
+//                case R.id.navigation_home:
+//                    mTextMessage.setText(R.string.title_home);
+//                    return true;
+//                case R.id.navigation_dashboard:
+//                    mTextMessage.setText(R.string.title_dashboard);
+//                    return true;
+//                case R.id.navigation_notifications:
+//                    mTextMessage.setText(R.string.title_notifications);
+//                    return true;
+//            }
+//            return false;
+//        }
+//
+//    };
+//
+//    @Override
+//    protected void onCreate(Bundle savedInstanceState) {
+//        super.onCreate(savedInstanceState);
+//        setContentView(R.layout.activity_main);
+//
+//        mTextMessage = (TextView) findViewById(R.id.message);
+//        BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.navigation);
+//        navigation.setOnNavigationItemSelectedListener(mOnNavigationItemSelectedListener);
+//    }
+//
+//}
