--- Mobile-Member/app/src/main/java/com/example/jayclarens/buttomnavigationviewfragmented/FragmentFilm.java
@@ -1,29 +1,0 @@
-package com.example.jayclarens.buttomnavigationviewfragmented;
-
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class FragmentFilm extends Fragment {
-
-
-    public FragmentFilm() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_film, container, false);
-    }
-
-}
