+++ Mobile-Member/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.jayclarens.buttomnavigationviewfragmented.MainActivity">
+    tools:context="com.TA.MVP.appmobilemember.MainActivity">
 
     <FrameLayout
         android:id="@+id/main_container"
