+++ Mobile-Member/app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.example.jayclarens.buttomnavigationviewfragmented.FragmentHome">
+             tools:context="com.TA.MVP.appmobilemember.FragmentHome">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
