+++ Mobile-Member/app/src/main/res/layout/fragment_musik.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.example.jayclarens.buttomnavigationviewfragmented.FragmentMusik">
+             tools:context="com.TA.MVP.appmobilemember.FragmentMusik">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
